improve pendingEOSE (commitdesc. avail.)
Normally, If it receives the first EOSE from one of existing relays, It
will forward the EOSE to client and "that's it", it ignores the further
EOSE messages. However, as the result, This confuses some nostr client
as it expects EOSE as "oh, that's the only events that i got", and there
we go something broke apart.

This commit, should fix at some cases, Especially for the one with NIP-50 supported relays (searching, etc)

Signed-off-by: Yonle <yonle@lecturify.net>