RELEASE: 2.0.8
Small and useful improvements and a little bit touch on moderations.

- Improvements on receiving new events in save mode (a.k.a. pause_on_limit)

Firewall / moderation things:
- A way to block an event owner's pubkey to have the event published
  into your bouncer. Useful to stop annoying actors to abuse your
  bouncer or in moderated nostr relay.

- Ability to block an IP address to connect into the bouncer's websocket.
  Useful to stop scrappers to constantly dig in your bouncer

I am not going to focus much on moderation however. So please bear in mind
that if you ask too much to add moderation stuffs into bostr, It will be
an different project. Because that's not what the goal of this project
for.

As of now i am still focusing to improve the save mode (pause_on_limit).

Signed-off-by: Yonle <yonle@lecturify.net>